pyproject.toml
[project]
name = "mail-management-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aioredis>=2.0.1",
    "asyncio>=4.0.0",
    "chromadb>=1.1.0",
    "fastapi>=0.116.1",
    "python-jose[cryptography]>=3.5.0",
    "requests>=2.32.5",
    "uvicorn>=0.35.0",
]
